requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
execnet==2.1.2
Flask==3.1.3
greenlet==3.3.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pillow==12.1.1
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.19.2
pytest==9.0.2
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
